Refactor Grid's file & folder structure
