+++ src/components/common/Grid/index.js
@@ -1,0 +1,3 @@
+export { default as Column } from './Column';
+export { default as Container } from './Container';
+export { default as Row } from './Row';
